Create Feedback Reports

make_feedback creates feedback files from a template .Rmd file

make_feedback(
  marks,
  template,
  filename = NULL,
  group_by = NULL,
  filter_by = NULL,
  quiet = FALSE,
  ...
)

Arguments

marks
data frame of marks or path to file

template
path to the .Rmd template file for feedback

filename
path to save files, defaults to "feedback/[group_by columns]", but you can reference any column names inside square brackets (e.g., "assignment_[moodle_id]_/[Student ID]")

group_by
columns to group by (e.g., if the marking file contains multiple rows per student) defaults to grouping by row if NULL

filter_by
optional vector for filtering, e.g. c(ID = 1) or c(question = "A")

quiet
print a message for each rendering

...
extra data to make available in the template file

Examples

if (FALSE) {
# saves each feedback file in a separate folder with the ID
make_feedback(demo_marks, "template.Rmd", "fb/[ID]/exam_Q[question]", "ID")
}